Moved functionality to set file metadata into files package
